+++ src/main/kotlin/dev/murek/elixirij/ExConfigurable.kt
@@ -10,9 +10,11 @@ import com.intellij.platform.lsp.api.LspServerManager
 import com.intellij.ui.components.JBTextField
 import com.intellij.ui.dsl.builder.*
 import dev.murek.elixirij.lsp.ExLspServerSupportProvider
+import dev.murek.elixirij.lsp.Expert
 import dev.murek.elixirij.toolchain.ExExecutableTestButton
 import dev.murek.elixirij.toolchain.ExToolchain
 import dev.murek.elixirij.toolchain.ExToolchainManager
+import javax.swing.JRadioButton
 
 class ExConfigurable(private val project: Project) : BoundConfigurable(
     ExBundle.message("configurable.elixir.displayName")
@@ -63,6 +65,8 @@ class ExConfigurable(private val project: Project) : BoundConfigurable(
         }
 
         group(ExBundle.message("configurable.expert.group.title")) {
+            lateinit var automaticModeRadio: Cell<JRadioButton>
+            lateinit var customModeRadio: Cell<JRadioButton>
             buttonsGroup {
                 row {
                     radioButton(ExBundle.message("configurable.expert.mode.disabled"), ExpertMode.DISABLED)
@@ -74,15 +78,40 @@ class ExConfigurable(private val project: Project) : BoundConfigurable(
                     radioButton(ExBundle.message("configurable.expert.mode.automatic"), ExpertMode.AUTOMATIC)
                         .applyToComponent {
                             name = "expertModeAutomaticRadio"
+                        }.also {
+                            automaticModeRadio = it
                         }
                 }
                 row {
                     radioButton(ExBundle.message("configurable.expert.mode.custom"), ExpertMode.CUSTOM)
                         .applyToComponent {
                             name = "expertModeCustomRadio"
+                        }.also {
+                            customModeRadio = it
                         }
                 }
             }.bind(settings::expertMode)
+
+            buttonsGroup {
+                row(ExBundle.message("configurable.expert.releaseChannel.label")) {
+                    radioButton(
+                        ExBundle.message("configurable.expert.releaseChannel.stable"),
+                        ExpertReleaseChannel.STABLE
+                    )
+                        .applyToComponent {
+                            name = "expertReleaseChannelStableRadio"
+                        }
+                    radioButton(
+                        ExBundle.message("configurable.expert.releaseChannel.nightly"),
+                        ExpertReleaseChannel.NIGHTLY
+                    )
+                        .applyToComponent {
+                            name = "expertReleaseChannelNightlyRadio"
+                        }
+                }
+            }
+                .bind(settings::expertReleaseChannel)
+                .enabledIf(automaticModeRadio.selected)
 
             row(ExBundle.message("configurable.expert.customPath.label")) {
                 textFieldWithBrowseButton(
@@ -95,11 +124,16 @@ class ExConfigurable(private val project: Project) : BoundConfigurable(
                     .bindText(settings::expertCustomExecutablePath.toNonNullableProperty(""))
                     .align(AlignX.FILL)
             }
+                .enabledIf(customModeRadio.selected)
         }
     }
 
     override fun apply() {
         super.apply()
+
+        if (settings.expertMode == ExpertMode.AUTOMATIC) {
+            Expert.getInstance(project).checkUpdates()
+        }
 
         LspServerManager.getInstance(project).stopAndRestartIfNeeded(ExLspServerSupportProvider::class.java)
     }